refactor: align serialization to 1.7.3, move to auth package and fix plugin configuration
